requirements.txt
aiofile==3.9.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.8
azure-common==1.1.28
azure-core==1.38.2
azure-identity==1.25.2
azure-mgmt-compute==37.2.0
azure-mgmt-core==1.6.0
azure-mgmt-resource==25.0.0
azure-mgmt-subscription==3.1.1
backports.tarfile==1.2.0
beartype==0.22.9
cachetools==7.0.1
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cyclopts==4.6.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastmcp==3.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.2
jsonschema-specifications==2025.9.1
keyring==25.7.0
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
msal==1.35.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
packaging==26.0
pathable==0.5.0
platformdirs==4.9.2
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyperclip==1.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
SecretStorage==3.5.0
sse-starlette==3.2.0
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
